# docs/swagger.yaml
definitions:
  config.SwaggerInfoResult:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.Product:
    properties:
      amount:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      updatedAt:
        type: string
    type: object
  product.createRequest:
    properties:
      amount:
        type: integer
      description:
        type: string
      name:
        type: string
      price:
        type: integer
    required:
    - amount
    - name
    - price
    type: object
info:
  contact: {}
  license: {}
paths:
  /v1/healthcheck:
    get:
      consumes:
      - application/json
      description: HealthCheck Service API
      parameters:
      - description: (en, th)
        in: header
        name: Accept-Language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/config.SwaggerInfoResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/config.SwaggerInfoResult'
      summary: HealthCheck
      tags:
      - Health
  /v1/products:
    get:
      consumes:
      - application/json
      description: Get Products Service API
      parameters:
      - description: (en, th)
        in: header
        name: Accept-Language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/config.SwaggerInfoResult'
      summary: Get Products
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Create Product Service API
      parameters:
      - description: (en, th)
        in: header
        name: Accept-Language
        type: string
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.createRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/config.SwaggerInfoResult'
      summary: Create Product
      tags:
      - Product
  /v1/products/{id}:
    get:
      consumes:
      - application/json
      description: Get Product Service API
      parameters:
      - description: (en, th)
        in: header
        name: Accept-Language
        type: string
      - description: query by product_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/config.SwaggerInfoResult'
      summary: Get Product
      tags:
      - Product
swagger: "2.0"